fix:サービスクラス、メインクラス 入力チェックを実装したため、記述を若干変更
入力チェックを実装した際に戻り値を全てObject型にしたため、呼出、戻り値を入れる際にキャストを呼出ごとに追加。

+++ refactoring_app_BesshoSora/src/jp/co/sss/crud/main/MainSystem.java
@@ -34,7 +34,7 @@ public class MainSystem {
 				ConsoleWriter.showMenu();
 
 				// メニュー番号の入力
-				menuNo = MenuNoReader.menuNoReader();
+				menuNo = (int) MenuNoReader.menuNoReader();
 
 				// 機能の呼出
 				switch (menuNo) {
