+++ refactoring_app_BesshoSora/src/jp/co/sss/crud/service/EmployeeDeleteService.java
@@ -27,7 +27,7 @@ public class EmployeeDeleteService {
 	 */
 	public static void deleteByEmpId() throws IllegalInputException, SystemErrorException {
 		// 削除する社員IDを入力
-		int empId = EmployeeEmpIdReader.employeeEmpIdReader(ConstantMsg.GUIDANCE_DELETE_BY_EMP_ID);
+		int empId = (int) EmployeeEmpIdReader.employeeEmpIdReader(ConstantMsg.GUIDANCE_DELETE_BY_EMP_ID);
 
 		//機能の呼出
 		int result = EmployeeDAO.deleteByEmpIdDAO(empId);
